# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clubhaus API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/teams/{team_id}/staff-reviews": {
      "get": {
        "tags": [
          "coaching"
        ],
        "summary": "List Staff Reviews",
        "operationId": "list_staff_reviews_api_v1_teams__team_id__staff_reviews_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StaffReviewResponse"
                  },
                  "title": "Response List Staff Reviews Api V1 Teams  Team Id  Staff Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "coaching"
        ],
        "summary": "Create Staff Review",
        "operationId": "create_staff_review_api_v1_teams__team_id__staff_reviews_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/staff-reviews/{review_id}/decision": {
      "post": {
        "tags": [
          "coaching"
        ],
        "summary": "Decide Staff Review",
        "operationId": "decide_staff_review_api_v1_staff_reviews__review_id__decision_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffReviewDecision"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StaffReviewCreate": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message"
          },
          "proposed_changes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Proposed Changes"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "message"
        ],
        "title": "StaffReviewCreate"
      },
      "StaffReviewDecision": {
        "properties": {
          "decision": {
            "type": "string",
            "pattern": "^(accepted|kept|rejected)$",
            "title": "Decision"
          },
          "note": {
            "type": "string",
            "minLength": 1,
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "note"
        ],
        "title": "StaffReviewDecision"
      },
      "StaffReviewResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team Id"
          },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "author_membership_id": {
            "type": "string",
            "format": "uuid",
            "title": "Author Membership Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "proposed_changes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Proposed Changes"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision"
          },
          "decision_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Note"
          },
          "decided_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided By Id"
          },
          "decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decided At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_id",
          "session_id",
          "author_membership_id",
          "message",
          "proposed_changes",
          "status",
          "decision",
          "decision_note",
          "decided_by_id",
          "decided_at",
          "created_at"
        ],
        "title": "StaffReviewResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}